+++ src/test/java/kitchenpos/study/MenuCascadeTest.java
@@ -4,11 +4,11 @@ import java.math.BigDecimal;
 import java.util.Arrays;
 import java.util.List;
 import kitchenpos.dao.MenuDao;
-import kitchenpos.dao.ProductDao;
+import kitchenpos.product.domain.ProductRepository;
 import kitchenpos.domain.Menu;
 import kitchenpos.domain.MenuProduct;
 import kitchenpos.domain.MenuProducts;
-import kitchenpos.domain.Product;
+import kitchenpos.product.domain.Product;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -20,7 +20,7 @@ import org.springframework.transaction.annotation.Transactional;
 public class MenuCascadeTest {
 
     @Autowired
-    ProductDao productDao;
+    ProductRepository productRepository;
 
     @Autowired
     MenuDao menuDao;
@@ -30,7 +30,7 @@ public class MenuCascadeTest {
         //given
         Product productA = new Product(null, "nameA", BigDecimal.valueOf(1000L));
         Product productB = new Product(null, "nameB", BigDecimal.valueOf(2000L));
-        productDao.saveAll(Arrays.asList(productA, productB));
+        productRepository.saveAll(Arrays.asList(productA, productB));
 
         MenuProduct menuProductA = new MenuProduct(null, null, productA.getId(), 10L);
         MenuProduct menuProductB = new MenuProduct(null, null, productB.getId(), 20L);
